+++ MapLoading.cpp
@@ -194,7 +194,7 @@ void ReadCellToSegment(API& DF, WorldSegment& segment, int CellX, int CellY, int
       for(uint32_t i=0; i<numVeins; i++){
 				//TODO: This will be fixed in dfHack at some point, but right now objects that arnt veins pass through as. So we filter on vtable
 				//if(veins[i].vtable != 9302320) only suitable for some versions
-        if(veins[i].type >= groundTypes.size())
+        if((uint32_t)veins[i].type >= groundTypes.size())
 					continue;
         uint16_t row = veins[i].assignment[ly];
         bool set = (row & (1 << lx)) != 0;
